Shove structs into protocol.rs
