+++ src/garfanzo.rs
@@ -1,23 +1,23 @@
 use std::io;
+use std::fs::File;
 use std::io::prelude::*;
 
 use std::sync::mpsc;
 use std::thread;
 
+pub mod protocol;
 pub mod gui;
 mod i2c;
 mod serial;
 
-#[derive(Clone, Debug)]
-pub enum Event {
-    I2C(i2c::I2CEvent),
-    Serial(serial::SerialEvent),
-    Gui(gui::GuiEvent),
-}
+use protocol::*;
 
-fn main() {
+fn main() -> std::io::Result<()> {
+    let mut file = File::create("/tmp/altctrl.serial")?;
+    file.write_all(b"Hello, world!")?;
     let stdin = io::stdin();
     for line in stdin.lock().lines() {
         println!("Input received: {}", line.unwrap());
     }
+    Ok(())
 }
