+++ src/i2c/mod.rs
@@ -2,28 +2,11 @@
 
 use std::sync::mpsc::Sender;
 
-use serde::{Serialize, Deserialize};
+//use serde::{Serialize, Deserialize};
 
 use crate::Event;
 
-#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
-pub enum Button {
-    B1,
-    B2,
-    B3,
-    B4,
-}
-
-#[derive(Clone, Debug)]
-pub enum I2CEvent {
-    On(Button),
-    Off(Button),
-}
-
-
-pub struct I2CStruct {
-    tx: Sender<Event>
-}
+use crate::protocol::*;
 
 pub fn initialize(tx: Sender<Event>) -> I2CStruct {
     I2CStruct {
